ARC: Add automake to the list of dependencies
Ubuntu package 'automake' is required to build OpenOCD (specifically to make 'bootstrap') but it was missing from ARC readme file. I guess that at some point automake was a dependency for a libtool, autoconfigure or build-essential, but as of Ubuntu 13.04 that is not the case, so I've added an explicit mention of this package to readme.

Signed-off-by: Anton Kolesov <akolesov@synopsys.com>