dontmerge: Rename filename with Cyrillic symbols
Configuration file for 1986бве1т microcontroller has Cyrillic symbols in it.
That confuses Cygwin on windows, which generates file with krakoziabras
(Russian term for this) instead of proper symbols. <sarcasm> Apparently Cygwin
developers consider Unicode to be too advanced technology to be utilized in
their product. </sarcasm> This patch renames this file to use Latin
equivalents for those characters.

Signed-off-by: Anton Kolesov <anton.kolesov@synopsys.com>