Arc: Fix error with writing BRK instruction
There was an error with handling 4-byte software breakpoints: in case of
little endian systems instruction should be written to memory in middle
endian format, but it was written in little endian like it is data. This
patch fixes this.

This patch is specific to 4-byte long BRK instructions created by Z0
packet with lenth argument 4. 2-byte long BRK_S breakpoints are not
affected.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>