Arc: More fixes for 32bit breakpoints
Previous commit was not that good, as it didn't added some required
things.

First is that after writing breakpoint instruction OpenOCD verifies that
instruction has been written. For this to work properly instruction
should be converted from middle endian.

Second is that when ARC OpenOCD unset breakpoints it checks if the code
hasn't been modified and doesn't removes breakpoint if there is already
some another instruction. To work properly this function should convert
instruction from middle endian.

Third is that when target has halted for yet un

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>