ARC: Introduce CHECK_RETVAL macro and use it
NDS32 implementation has a nice CHECK_RETVAL macro which auto-checks for
return value of functions and returns if it is an error. This simplifies
code of functions calls, since we don't need to check for return value
explicitly, other than we have to wrap function call in this macro. This
macro doesn't help with function functions which allocate memory since it
returns immediately and allocated memory will not get freed. So in those
functions we have to check for return value manually. At least for now.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>