arc: Add an artificial delay when reading from DDR memory
When reading more than one word subsequently from DDR memory there was a problem that
second and further words were read as 0, instead of a real content. That
could have been easily reproduced with:

mww 0x10000000 0xde
mww 0x10000004 0xad
mdw 0x10000004 1
mdw 0x10000000 2

For the second read we would get "0x000000de 0x00000000". One way to fix
that is to reduce JTAG frequency. For example with EM Starter Kit EM7D@30MHz
error was resolved with JTAG frequency 2.2MHz. That increased time of bulk
memory reads (100 000 consecutive 32bit words) in 3.1 times plus it affects
all of the other JTAG operations.  Another solution (implemented in this
patch) is to write address of a memory word to read each time before reading
it. In general that is unnecessary for consecutive reads, since address is
auto-incremented, so here it is just an artificial way to add a required
delay. In comparison to reduced JTAG speed this patch decreases performance
of the same bulk memory read in 2.7 times, and it affects only memory reads
- all other JTAG operations are done at the same pace as before. In addition
using functionality added in previous patches memory layer applies this
delay only to the non-CCM memories, so ICCM/DCCM are not affected by this
patch as well.

I do have a feeling that this might be a bug in EM7D JTAG interface, but I
don't really know. See Synopsys STARs 9000830091 and 9000832538 for details.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>