arc: Add JTAG STATUS.FL checks after each operation
Check for FL bit in JTAG STATUS register and return error if this bit is
set, which usually indicates some kind of an error.

This however is disabled by default because it is known to break JTAG: there
is an undesired extra write after the indented one, so if we write, say,
DEBUG AUX register at 0x5, then we will also write same value to PC register
at 0x6, which is totally undesired.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>