arc: Do not halt target when examining
Target examination was halting the target. That is not expected by the
OpenOCD, target should be left in the same state as before examination. JTAG
works fine with unhalted targets, though of course there is a chance that
target state will change behind the scenes with unhalted target.

It seems that we also shouln't exit quickly from target_halt() in this case,
because target might be reset, but not halted.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>