arc: Support for 'general' registers
Add an option `-g` to `arc add-reg` command that specifies that registers is
"general". That means that register is ubiquitous and should be shown by GDB
command `info registers` (without extra arguments). Thus to optimize access
to those registers OpenOCD will always read those "general" registers in
save_context and will include them in g/G-packet. The "non-general"
registers will be lazy-read when required and have to be requested by GDB
explicitly via p-packet. Note, that because g/G-packet is continuous if
there is a "non-general" register between two "general" that that register
will also be read by save_context  and will be in g-packet. Thus effectively
any register before the last "general" register will be "general" as well.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>