arc: Set current target in event handlers
Handlers of core events should do actions on the core that generated event.
However OpenOCD doesn't set implicit "current target" to the core that
is generating an event and there is no way for us to figure out what was
the core that generated event. To workaround this problem with multicore
chips it is required that direct event handler should be not a generic proc,
but rather a local string that will pass target name as an argument to the
handling proc.

Note that in general case such handler should be written in a
double quotes, not curly braces, because in TCL double quotes are expanded
at the moment of evaluation (when we assign handler), while curly braces
would be expanded at the moment of execution, thus _TARGETNMAE would point
to the target current to moment of handler execution, something that we
are deliberately trying to avoid.

This patch also allows "arc_common_reset" to be called without arguments,
then rest will be done for current target. That should work fine for single
core systems, since there current target is always correct.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>