arc: Add TCL description for ARC HS targets
This separates TCL description of ARC v2 cores into EM and HS separate
parts. That is required because there are some incompatible differences
between them and current API doesn't allow to update required pieces
dynamically from TCL (although is only because required functionality is not
exposed to TCL, in general that is doable). So for proper functioning it is
required to include proper core-specific TCL file, although it also should
work with a more generic v2 TCL file, although it will not detect features
that are specific to a particular core.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>