arc: Introduce target types arc600, arc700 and arcv2
ARC used to have only arc32 target type. However there is also a need to
distinguish ARC 600 from rest of cores, because ARC 600 has DEBUG.SS bit.
Previous ARC used to distinguish ARC 600 from the rest by the target name,
however that is really wrong, target name is just a name, which should
represent core role in SoC, not its architecture. Architecture is
represented by thet target type.

Therefore this commits splits arc32 into several targets: arc600, arc700 and
arcv2. Difference is very-very small, so I even feel that such division is
an overkill. The only difference is that for arc600 DEBUG.SS bit will be set
when single instruction stepping. Function that doesn't this instruction
stepping is left intact, I don't want to separate it into several, due to
amount of shared code for different cores, so it has just an `if` for ARC
600. As a result arc600 target would just set a special flag. arc700 and
arcv2 are really identical, but I felt that it would be sematically wrong to
have arc600 and a unified arc700+arcv2, say arc32.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>