arc: Do not wait for JTAG operation success by default
One of the previous commits added an extra check to JTAG layer - after
doing register writes and reads OpenOCD now was polling JTAG STATUS
register. However it has been found that this check triggers a bug in some
ARC core implementations: EMs in EM Starter Kit, ARC 600 and ARC 700 in
AXS101. No bug is observed with EM6 in AXS101 or with HS in AXS102. The bug
is that when there is a write transaction and JTAG STATUS register is read
to check if write finished - there is an undesired second write at the next
register/memory word. That seems like a bug in the JTAG unit, in part because
what is written depends on whether first transaction was write to register
or memory. When writing to a register, the same value will be written to the
following register. When writing to memory, an *address* of previous write
destination would be written to the next word of memory. So if there is a
write to 0x8000_0000, then 0x8000_0000 would be written to 0x8000_0004.

As a result using "wait-until-write-is-finished on" is not possible with
a significant amount of core implementations. On the other hand it is seems
like a useful feature in some scenarios. But, apparently, cases where it is
needed are more rare than cases where it breaks writes.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>