arc: Fix invalid assert in arc_mem_is_slow_memory
In the arc_mem.c "arc_mem_is_slow_memory", there is this assert:
`assert(addr_end >= addr)` However, it will falsely trigger when checking the
end of 32 bit space. For example if

	addr = 0xFFFFF000
	size = 4
	count = 1024

Then the assertion will trigger as then addr_end is evaluated as 0.

Since addr_end points to the the address _after_ the end of memory region,
assert condition should be updated.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>