arc: Flush L2 cache at reset for ARC HS
ARC port of OpenOCD currently doesn't support L2 cache - it is not flushed
and invalidated when memory is being modified.  However that is not needed
for a simplified case where OpenOCD is needed only to do an initial load of
image into target and then let it run. Still, to be sure that this operation
is safe it is preferably to flush and invalidate L2 cache before loading
image. This can be easily done in the TCL configuration script, for example
in in reset handler.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>