rtos/Zephyr: read registers from correct location
Almost all architectures store the stack pointer in the _callee_saved
structure and we know which other registers are stored there as well.
The rest of the general registers should have been stored to the stack
of the thread. The only exception is x86-64, which is not yet in a
usable state and eventually has to store the stack pointer somewhere.

For ARM Cortex-M we must also skip the floating point registers on the
stack if they have been stored by the cpu and undo the stack pointer
realignment to 8 bytes. It is not possible to return the floating point
registers to GDB with the current OpenOCD RTOS API.

Fixes #4

Signed-off-by: Daniel Glöckner <dg@emlix.com>