rtos/Zephyr: support recent Zephyr versions
The current version reported by Zephyr in the offsets array is 1.
It is compatible to version 0 but offers a few more values in the
array. zephyrproject-rtos/zephyr#13461 adds a separate symbol to
hold the size of the array, so that future compatible additions
do not require the version number to be incremented.

Fixes #3

Signed-off-by: Daniel Glöckner <dg@emlix.com>