[RFC] target/cortex_m: Block extra srst asserts in target code
This is intended to be a temporary work around for an issue where
TCL scripts are toggling srst, and then the Cortex M target code
is again asserting srst. On the TI LPRF (CC26xx/CC13xx) devices
this results in an undebuggable state because the target code
does not restore the JTAG connection. LPRF targets come out of
reset in 2-pin mode and also require the ICEPick JTAG router to
put the Cortex M core into the scan chain. Neither are happening
after the Cortex M target code asserts srst.

Added a flag to the JTAG reset config to signal to the Cortex M
routine to -not- assert srst again. srst was already handled prior,
and the target has been reset and ready to debug.

Ideally, what really needs to happen for the LPRF targets is the
following.

For reset halt or reset init:
assert srst
run TCK for 50 ms
deassert srst
run TCK for 50 ms
switch to 4-pin mode
command ICEPick to put Cortex M TAP into scan chain
continue reconnecting debug
halt the core

For reset run:
assert srst
delay
deassert srst
delay
switch to 4-pin mode
command ICEPick to put Cortex M TAP into scan chain
continue reconnecting debug
leave core running

The running TCK periods during the srst assert/deassert on
these device is to trigger the LPRF's HIB (halt in boot)
mode where the target will stop execution at a hook in the
boot ROM so that debug can start at the very start of the
user application. The XDS110 code today already just does
this every time, but ideally this should be configurable
so that other probes can trigger HIB.

And note that debug on LPRF is not very stable without
using srst reset. It is common for user applications to
get the device in a mode that cannot be debugged unless
srst and HIB are used.

Change-Id: I424e6f66cfeee58ba17991004f6c8fd184af8099
Signed-off-by: Edward Fewell <efewell@ti.com>