cfg: Set STM32F746G discovery board to use 4000 kHz on reset-start
Currently board/stm32f746g-disco.cfg already uses the maximum adapter speed
available for reset-init events (4000 kHz). However it doesn't set that
speed for reset-start events and so the speed for that event is inherited
from the included config file arget/stm32f7x.cfg, which sets that speed to
just 2000 kHz.

That commit overrides the inherited speed for reset-start events setting
the adapter speed to be equal to the one used for reset-init events, i.e.
to its maximum speed (4000 kHz).

That change avoids the following messages to appear when fashing the
board:

Info : Unable to match requested speed 2000 kHz, using 1800 kHz
Info : Unable to match requested speed 2000 kHz, using 1800 kHz

It also increases the throughput a bit when flashing the device.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>