Revert "target: reset target examined flag if target::examine() fails"
This reverts commit 98d9f1168cbdc59e4c2c0b1f01b225a4df9ad98a because
it causes flashing issues with some Zephyr targets (notably, STM32
family devices).

The patch in itself does not seem to be doing anything wrong in
particular; but, it uncovers various problems with the Zephyr-side
OpenOCD configurations -- mainly, not resetting the target device on
debugger connect, which may lead to connection failures because Zephyr
puts the target CPU to sleep while idling and the CPU cannot respond to
the debugger's request in this state.

For more details, refer to the GitHub issue
zephyrproject-rtos/zephyr#50590.

Revert this commit once the Zephyr-side OpenOCD configurations are
fixed.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>