flash: nor: efm32: align switch and case statements
The coding style requires the 'case' to be at the same indentation
level of its 'switch' statement.

Align the code accordingly.

This patch is a subset of the commit 36a2b6c6f ("flash: nor: align
switch and case statements"). While the original patch changed many
drivers, this patch only apply the changes on src/flash/nor/efm32.c.
This is required to avoid conflict in the forther patches we want to
backport.

Change-Id: I6be44efd5189b671caabcf6753bb82ef44521440
Reviewed-on: https://review.openocd.org/c/openocd/+/9034
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>